Moved 'update piece' functionality to game namespace. Functionality should now be accessible via games/pieces/:id
