+++ app/helpers/games/pieces_helper.rb
@@ -1,0 +1,2 @@
+module Games::PiecesHelper
+end
