--- app/views/pieces/show.html.erb
@@ -1,107 +1,0 @@
-<div class="chessboard">
-  <!-- 1st -->
-  <div id="a8" class="white"></div>
-  <div id="b8" class="black"></div>
-  <div id="c8" class="white"></div>
-  <div id="d8" class="black"></div>
-  <div id="e8" class="white"></div>
-  <div id="f8" class="black"></div>
-  <div id="g8" class="white"></div>
-  <div id="h8" class="black"></div>
-  <!-- 2nd -->
-  <div id="a7" class="black"></div>
-  <div id="b7" class="white"></div>
-  <div id="c7" class="black"></div>
-  <div id="d7" class="white"></div>
-  <div id="e7" class="black"></div>
-  <div id="f7" class="white"></div>
-  <div id="g7" class="black"></div>
-  <div id="h7" class="white"></div>
-  <!-- 3th -->
-  <div id="a6" class="white"></div>
-  <div id="b6" class="black"></div>
-  <div id="c6" class="white"></div>
-  <div id="d6" class="black"></div>
-  <div id="e6" class="white"></div>
-  <div id="f6" class="black"></div>
-  <div id="g6" class="white"></div>
-  <div id="h6" class="black"></div>
-  <!-- 4st -->
-  <div id="a5" class="black"></div>
-  <div id="b5" class="white"></div>
-  <div id="c5" class="black"></div>
-  <div id="d5" class="white"></div>
-  <div id="e5" class="black"></div>
-  <div id="f5" class="white"></div>
-  <div id="g5" class="black"></div>
-  <div id="h5" class="white"></div>
-  <!-- 5th -->
-  <div id="a4" class="white"></div>
-  <div id="b4" class="black"></div>
-  <div id="c4" class="white"></div>
-  <div id="d4" class="black"></div>
-  <div id="e4" class="white"></div>
-  <div id="f4" class="black"></div>
-  <div id="g4" class="white"></div>
-  <div id="h4" class="black"></div>
-  <!-- 6th -->
-  <div id="a3" class="black"></div>
-  <div id="b3" class="white"></div>
-  <div id="c3" class="black"></div>
-  <div id="d3" class="white"></div>
-  <div id="e3" class="black"></div>
-  <div id="f3" class="white"></div>
-  <div id="g3" class="black"></div>
-  <div id="h3" class="white"></div>
-  <!-- 7th -->
-  <div id="a2" class="white"></div>
-  <div id="b2" class="black"></div>
-  <div id="c2" class="white"></div>
-  <div id="d2" class="black"></div>
-  <div id="e2" class="white"></div>
-  <div id="f2" class="black"></div>
-  <div id="g2" class="white"></div>
-  <div id="h2" class="black"></div>
-  <!-- 8th -->
-  <div id="a1" class="black"></div>
-  <div id="b1" class="white"></div>
-  <div id="c1" class="black"></div>
-  <div id="d1" class="white"></div>
-  <div id="e1" class="black"></div>
-  <div id="f1" class="white"></div>
-  <div id="g1" class="black"></div>
-  <div id="h1" class="white"></div>
-</div>
-<script>
-  $(function() {
-    getDiv(<%= @piece.location_x %>,<%= @piece.location_y %>);
-    function getDiv(x,y) {
-      var new_x = "";
-      switch (x) {
-        case 1: new_x = "a";
-        break;
-        case 2: new_x = "b";
-        break;
-        case 3: new_x = "c";
-        break;
-        case 4: new_x = "d";
-        break;
-        case 5: new_x = "e";
-        break;
-        case 6: new_x = "f";
-        break;
-        case 7: new_x = "g";
-        break;
-        case 8: new_x = "h";
-        break;
-      };
-      var divSelector = '#' + new_x + y;
-      var currentClass = $(divSelector).attr('class');
-      if (currentClass = "white") {
-        $(divSelector).addClass("highlighted-white");  
-      } else {
-        $(divSelector).addClass("highlighted-black");  
-      }
-    }
-  });
-</script>
