+++ config/routes.rb
@@ -1,7 +1,8 @@
  Rails.application.routes.draw do
   devise_for :users, controllers: { omniauth_callbacks: 'users/omniauth_callbacks' }
   root 'static_pages#index'
-  resources :games do
+  resources :games
+  namespace :games do
     resources :pieces, only: [:show, :update]
   end
 end
