+++ spec/controllers/games/pieces_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Games::PiecesController, type: :controller do
+
+end
